requirements.txt
alembic==1.8.1
certifi==2022.6.15
click==8.0.4
Flask==2.1.3
Flask-Injector==0.14.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
injector==0.20.1
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
pip==22.1.2
pyparsing==3.0.9
python-dotenv==0.20.0
setuptools==61.2.0
six==1.16.0
SQLAlchemy==1.4.40
Werkzeug==2.0.3
wheel==0.37.1
